pyproject.toml
[project]
name = "conditional-misalignment"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "llmcomp>=1.2.3",
    "matplotlib>=3.10.0",
    "numpy>=1.24.0",
    "openai>=1.0.0",
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "reportlab>=4.0.0",
]
